refactor: common & Widget 폴더 정리
공용 컴포넌트: Common & 위젯 컴포넌트: Widget 폴더로 이동
