+++ src/components/index.ts
@@ -1,5 +1,5 @@
 import Background from './Background';
-import Clock from './Clock';
-import Weather from './Weather';
+import Clock from './Widget/Clock';
+import Weather from './Widget/Weather/Weather';
 
 export { Background, Clock, Weather };
